resolve: `lifetimes.rs` -> `late/lifetime.rs`
